Canonical Spin Systems

ExactDiagonalization.EDCore.EDKind — Method

EDKind(::Type{<:Hilbert{<:Spin}})

The kind of the exact diagonalization method applied to a canonical quantum spin lattice system.

QuantumLattices.DegreesOfFreedom.Metric — Method

Metric(::EDKind{:SED}, ::Hilbert{<:Spin}) -> OperatorUnitToTuple

Get the index-to-tuple metric for a canonical quantum spin lattice system.