ExactDiagonalization

ExactDiagonalization

Implements exact diagonalization

ExactDiagonalization.HilbertSpace — Type.

HilbertSpace{QN}

Abstract Hilbert space with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> spin_site = Site{Int64}([State{Int64}("Up", +1), State{Int64}("Dn", -1)])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

julia> hs = HilbertSpace{Int64}([spin_site, spin_site])
HilbertSpace{Int64}(Site{Int64}[Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)]), Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])], [1, 1], [0, 1, 2])

source

ExactDiagonalization.Site — Type.

Site{QN}

A site with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> up = State{Int}("Up", 1); dn = State{Int}("Dn", -1);

julia> Site([up, dn])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

source

ExactDiagonalization.SparseState — Type.

struct SparseState{Scalar<:Number, BR}

Represents a row vector?

source

ExactDiagonalization.State — Type.

State{QN}

State with quantum number type QN.

Examples

julia> using ExactDiagonalization, StaticArrays

julia> up = State{Int}("Up", 1)
State{Int64}("Up", 1)

julia> State("Dn", SVector{2, Int}([-1, 1]))
State{SArray{Tuple{2},Int64,1,2}}("Dn", [-1, 1]) 

source

ExactDiagonalization.apply! — Method.

apply!

Apply operator to psi and add it to out.

source

ExactDiagonalization.bitwidth — Method.

bitwidth(site ::Site)

Number of bits necessary to represent the states of the given site.

source

ExactDiagonalization.compress — Method.

Convert an array of indices (of states) to binary representation

source

ExactDiagonalization.dimension — Method.

dimension

Dimension of the Concrete Hilbert space, i.e. number of basis vectors.

source

ExactDiagonalization.dimension — Method.

dimension(site ::Site)

Hilbert space dimension of a given site ( = number of states).

source

ExactDiagonalization.extract — Method.

Convert binary representation to an array of indices (of states)

Examples ≡≡≡≡≡≡≡≡≡≡

source

ExactDiagonalization.get_state — Method.

get_state(site ::Site{QN}, binrep ::BR) where {QN, BR<:Unsigned}

Returns the state of site represented by the bits binrep.

source

ExactDiagonalization.is_compatible — Method.

is_compatible

Check whether the fractional momentum ([0, 1)ᴺ) compatible with the identity translation. i.e. k¹ R¹ + k² R² + ... + kᴺ Rᴺ = 0 (mod 1)

Arguments

fractional_momentum ::AbstractVector{Rational} : k

identity_translation ::AbstractVector{<:Integer} : R

source

ExactDiagonalization.realize — Method.

realize(hs; BR ::DataType=UInt)

Make a HilbertSpaceRealization with all the basis vectors of the specified HilbertSpace.

Arguments

hs ::HilbertSpace{QN}: Abstract Hilbert space

BR ::DataType=UInt: Binary representation type

source

ExactDiagonalization.realize — Method.

realize(hs; BR ::DataType=UInt)

Make a HilbertSpaceRealization with all the basis vectors of the specified HilbertSpace.

Arguments

hs ::HilbertSpace{QN}: Abstract Hilbert space

allowed: Allowed quantum numbers

BR ::DataType=UInt: Binary representation type

source