ExactDiagonalization

Julia package for the exact diagonalization method in condensed matter physics.

Based on the symbolic operator representation of a quantum lattice system in condensed matter physics that is generated by the package QuantumLattices, exact diagonalization method is implemented for fermionic, hard-core-bosonic and spin systems.

Installation

In Julia v1.6+, please type ] in the REPL to use the package mode, then type this command:

pkg> add ExactDiagonalization

Getting Started

Examples of exact diagonalization method for quantum lattice system

Python counterpart

HamiltonianPy: in fact, the authors of this Julia package worked on the python package at first and only turned to Julia later.

Manuals

ExactDiagonalization.BinaryBases — Type

BinaryBases{B<:BinaryBasis, T<:AbstractVector{B}} <: Sector

A set of binary bases.

ExactDiagonalization.BinaryBases — Method

BinaryBases(nstate::Integer, nparticle::Integer)
BinaryBases(states, nparticle::Integer)

Construct a set of binary bases that preserves the particle number conservation.

ExactDiagonalization.BinaryBases — Method

BinaryBases(nstate::Integer)
BinaryBases(states)

Construct a set of binary bases that does not preserve the particle number conservation.

ExactDiagonalization.BinaryBasis — Type

BinaryBasis{I<:Unsigned}

Binary basis represented by an unsigned integer.

ExactDiagonalization.BinaryBasis — Method

BinaryBasis(states; filter=index->true)
BinaryBasis{I}(states; filter=index->true) where {I<:Unsigned}

Construct a binary basis with the given occupied orbitals.

ExactDiagonalization.BinaryBasisRange — Type

BinaryBasisRange{I<:Unsigned} <: VectorSpace{BinaryBasis{I}}

A continuous range of binary basis.

ExactDiagonalization.ED — Type

ED{L<:AbstractLattice, G<:Generator, M<:Image} <: Engine

Exact diagonalization method of a quantum lattice system.

ExactDiagonalization.ED — Method

ED(lattice::AbstractLattice, hilbert::Hilbert, terms::Tuple{Vararg{Term}}, targetspace::TargetSpace; boundary::Boundary=plain)

Construct the exact diagonalization method for a quantum lattice system.

ExactDiagonalization.EDKind — Type

EDKind{K}

The kind of the exact diagonalization method applied to a quantum lattice system.

ExactDiagonalization.EDMatrix — Type

EDMatrix{S<:Sector, M<:SparseMatrixCSC} <: OperatorPack{M, Tuple{S, S}}

Matrix representation of quantum operators between a ket and a bra Hilbert space.

ExactDiagonalization.EDMatrix — Method

EDMatrix(sector::Sector, m::SparseMatrixCSC)

Construct a matrix representation when the ket and bra spaces share the same bases.

ExactDiagonalization.EDMatrixRepresentation — Type

EDMatrixRepresentation{S<:Sector, T} <: MatrixRepresentation

Exact matrix representation of a quantum lattice system on a target Hilbert space.

ExactDiagonalization.EDMatrixRepresentation — Method

EDMatrixRepresentation(target::TargetSpace, table)

Construct a exact matrix representation.

ExactDiagonalization.Sector — Type

abstract type Sector <: OperatorUnit

A sector of the Hilbert space which form the bases of an irreducible representation of the Hamiltonian of a quantum lattice system.

ExactDiagonalization.SectorFilter — Type

SectorFilter{S} <: Transformation

Filter the target bra and ket Hilbert spaces.

ExactDiagonalization.TargetSpace — Type

TargetSpace{S<:Sector} <: VectorSpace{S}

The target Hilbert space in which the exact diagonalization method is performed, which could be the direct sum of several sectors.

ExactDiagonalization.TargetSpace — Method

TargetSpace(sector::Sector, sectors::Sector...)

Construct a target space from sectors.

QuantumLattices.Essentials.DegreesOfFreedom.Metric — Method

Metric(::EDKind{:FED}, hilbert::Hilbert{<:Fock}) -> OIDToTuple
Metric(::EDKind{:SED}, hilbert::Hilbert{<:Spin}) -> OIDToTuple

Get the oid-to-tuple metric for a free fermionic/bosonic system or a free phononic system.

Base.count — Method

count(basis::BinaryBasis) -> Int
count(basis::BinaryBasis, start::Integer, stop::Integer) -> Int

Count the number of occupied single-particle states.

Base.isone — Method

isone(basis::BinaryBasis, state::Integer) -> Bool

Judge whether the specified single-particle state is occupied for a basis.

Base.iszero — Method

iszero(basis::BinaryBasis, state::Integer) -> Bool

Judge whether the specified single-particle state is unoccupied for a basis.

Base.iterate — Function

iterate(basis::BinaryBasis, state=nothing)

Iterate over the numbers of the occupied single-particle orbitals.

Base.one — Method

one(basis::BinaryBasis, state::Integer) -> BinaryBasis

Get a new basis with the specified single-particle state occupied.

Base.zero — Method

zero(basis::BinaryBasis, state::Integer) -> BinaryBasis

Get a new basis with the specified single-particle state unoccupied.

ExactDiagonalization.productable — Method

productable(bs₁::BinaryBases, bs₂::BinaryBases) -> Bool

Judge whether two sets of binary bases could be direct producted.

ExactDiagonalization.sumable — Method

sumable(bs₁::BinaryBases, bs₂::BinaryBases) -> Bool

Judge whether two sets of binary bases could be direct summed.

Strictly speaking, two sets of binary bases could be direct summed if and only if they have no intersection. The time complexity to check the intersection is O(n log n), which costs a lot when the dimension of the binary bases is huge. It is also possible to judge whether they could be direct summed by close investigations on their ids, i.e. the single-particle states and occupation number. It turns out that this is a multi-variable pure integer linear programming problem. In the future, this function would be implemented based on this observation. At present, the direct summability should be handled by the users in priori.

LinearAlgebra.eigen — Method

eigen(m::EDMatrix; nev=6, which=:SR, tol=0.0, maxiter=300, sigma=nothing, v₀=dtype(m)[]) -> Eigen

Solve the eigen problem by the restarted Lanczos method provided by the Arpack package.

QuantumLattices.Essentials.QuantumOperators.matrix — Method

matrix(ed::ED, sector::Sector; kwargs...) -> EDMatrix
matrix(ed::ED, sector=first(ed.Hₘ.transformation.brakets); kwargs...) -> EDMatrix

Get the sparse matrix representation of a quantum lattice system in a sector of the target space.

QuantumLattices.Essentials.QuantumOperators.matrix — Method

matrix(op::Operator, braket::NTuple{2, BinaryBases}, table; dtype=valtype(op)) -> SparseMatrixCSC{dtype, Int}
matrix(ops::Operators, braket::NTuple{2, BinaryBases}, table; dtype=valtype(eltype(ops))) -> SparseMatrixCSC{dtype, Int}

Get the CSC-formed sparse matrix representation of an operator.

Here, table specifies the order of the operator ids.

QuantumLattices.Interfaces.:⊕ — Method

⊕(sector::Sector, sectors::Union{Sector, TargetSpace}...) -> TargetSpace
⊕(target::TargetSpace, sectors::Union{Sector, TargetSpace}...) -> TargetSpace

Get the direct sum of sectors and target spaces.

QuantumLattices.Interfaces.:⊗ — Method

⊗(bs₁::BinaryBases, bs₂::BinaryBases) -> BinaryBases

Get the direct product of two sets of binary bases.

QuantumLattices.Interfaces.:⊗ — Method

⊗(basis₁::BinaryBasis, basis₂::BinaryBasis) -> BinaryBasis

Get the direct product of two binary bases.