API

API

ExactDiagonalization.HilbertSpace — Type.

HilbertSpace{QN}

Abstract Hilbert space with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> spin_site = Site{Int64}([State{Int64}("Up", +1), State{Int64}("Dn", -1)])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

julia> hs = HilbertSpace{Int64}([spin_site, spin_site])
HilbertSpace{Int64}(Site{Int64}[Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)]), Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])], [1, 1], [0, 1, 2])

source

ExactDiagonalization.Permutation — Type.

Permutation(perms ::AbstractVector{Int}; max_cycle=2048)

Create a permutation of integers from 1 to n. perms should be a permutation of 1:n.

Arguments

perms: an integer vector containing a permutation of integers from 1 to n

max_cycle: maximum length of cycle

source

ExactDiagonalization.Site — Type.

Site{QN}

A site with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> up = State{Int}("Up", 1); dn = State{Int}("Dn", -1);

julia> Site([up, dn])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

source

ExactDiagonalization.SparseState — Type.

struct SparseState{Scalar<:Number, BR}

Represents a row vector. Free.

source

ExactDiagonalization.State — Type.

State{QN}

State with quantum number type QN.

Examples

julia> using ExactDiagonalization, StaticArrays

julia> up = State{Int}("Up", 1)
State{Int64}("Up", 1)

julia> State("Dn", SVector{2, Int}([-1, 1]))
State{SArray{Tuple{2},Int64,1,2}}("Dn", [-1, 1])

source

ExactDiagonalization.apply! — Method.

apply!

Apply operator to psi and add it to out.

source

ExactDiagonalization.apply! — Method.

apply!(out, opr, state; range=1:size(opr, 1))

Perform out += opr * state. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Call apply_serial! if Threads.nthreads() == 1, and apply_parallel! if greater.

Arguments

out ::Vector{S1}

state ::AbstractVector{S2}

opr ::OperatorRepresentation{HSR, O}

range ::AbstractVector{<:Integer}=1:size(opr, 1)

source

ExactDiagonalization.apply! — Method.

apply!(out, opr, state; range=1:size(opr, 2))

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Call apply_serial! if Threads.nthreads() == 1, and apply_parallel! if greater.

Arguments

out ::Vector{S1}

opr ::OperatorRepresentation{HSR, O}

state ::AbstractVector{S2}

range ::AbstractVector{<:Integer}=1:size(opr, 2)

source

ExactDiagonalization.apply_parallel! — Method.

apply_parallel!(out, state, opr; range=1:size(opr, 1))

Perform out += state * opr. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Multi-threaded version.

Arguments

out ::Vector{S1}

state ::AbstractVector{S2}

opr ::OperatorRepresentation{HSR, O}

range ::AbstractVector{<:Integer}=1:dimension(opr.hilbert_space_representation)

source

ExactDiagonalization.apply_parallel! — Method.

apply_parallel!(out, opr, state; range=1:size(opr, 2))

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Multi-threaded version.

Arguments

out ::Vector{S1}

opr ::OperatorRepresentation{HSR, O}

state ::AbstractVector{S2}

range ::AbstractVector{<:Integer}=1:size(opr, 2)

source

ExactDiagonalization.apply_serial! — Method.

apply_serial!(out, state, opr; range=1:size(opr, 1))

Perform out += state * opr. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Single-threaded version.

Arguments

out ::Vector{S1}

state ::AbstractVector{S2}

opr ::OperatorRepresentation{HSR, O}

range ::AbstractVector{<:Integer}=1:size(opr, 1)

source

ExactDiagonalization.apply_serial! — Method.

apply_serial!(out, opr, state; range=1:size(opr, 2))

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Single-threaded version.

Arguments

out ::Vector{S1}

opr ::OperatorRepresentation{HSR, O}

state ::AbstractVector{S2}

range ::AbstractVector{<:Integer}=1:size(opr, 2)

source

ExactDiagonalization.bitwidth — Method.

Total number of bits

julia> using ExactDiagonalization

julia> spin_site = Site{Int64}([State{Int64}("Up", +1), State{Int64}("Dn", -1)])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

julia> hs = HilbertSpace{Int64}([spin_site, spin_site, spin_site,])
HilbertSpace{Int64}(Site{Int64}[Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)]), Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)]), Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])], [1, 1, 1], [0, 1, 2, 3])

julia> bitwidth(hs)
3

source

ExactDiagonalization.bitwidth — Method.

bitwidth(site ::Site)

Number of bits necessary to represent the states of the given site.

source

ExactDiagonalization.dimension — Method.

dimension

Dimension of the Concrete Hilbert space, i.e. number of basis vectors.

source

ExactDiagonalization.dimension — Method.

dimension(site ::Site)

Hilbert space dimension of a given site ( = number of states).

source

ExactDiagonalization.extract — Method.

Convert binary representation to an array of indices (of states)

Examples ≡≡≡≡≡≡≡≡≡≡

source

ExactDiagonalization.get_quantum_number — Method.

get_quantum_number

source

ExactDiagonalization.get_row_iterator — Method.

May contain duplicates

source

ExactDiagonalization.get_state — Method.

get_state(site ::Site{QN}, binrep ::BR) where {QN, BR<:Unsigned}

Returns the state of site represented by the bits binrep.

source

ExactDiagonalization.is_compatible — Method.

is_compatible

Check whether the fractional momentum ([0, 1)ᴺ) compatible with the identity translation. i.e. k¹ R¹ + k² R² + ... + kᴺ Rᴺ = 0 (mod 1)

Arguments

fractional_momentum ::AbstractVector{Rational} : k

identity_translation ::AbstractVector{<:Integer} : R

source

ExactDiagonalization.qntype — Method.

qntype

source

ExactDiagonalization.quantum_number_sectors — Method.

quantum_number_sectors

source

ExactDiagonalization.represent — Method.

represent(hs; BR ::DataType=UInt)

Make a HilbertSpaceRepresentation with all the basis vectors of the specified HilbertSpace.

Arguments

hs ::HilbertSpace{QN}: Abstract Hilbert space

allowed: Allowed quantum numbers

BR ::DataType=UInt: Binary representation type

source

ExactDiagonalization.represent — Method.

represent(hs; BR ::DataType=UInt)

Make a HilbertSpaceRepresentation with all the basis vectors of the specified HilbertSpace.

Arguments

hs ::HilbertSpace{QN}: Abstract Hilbert space

BR ::DataType=UInt: Binary representation type

source

ExactDiagonalization.symmetry_reduce — Method.

source

ExactDiagonalization.symmetry_reduce_serial — Method.

symmetry_reduce_serial(hsr, trans_group, frac_momentum; ComplexType=ComplexF64, tol=sqrt(eps(Float64)))

Symmetry-reduce the HilbertSpaceRepresentation using translation group.

source

Base.:* — Method.

*(p1 ::Permutation, p2 ::Permutation)

Multiply the two permutation. Return [p2.map[x] for x in p1.map].

Examples

julia> using ExactDiagonalization

julia> Permutation([1,3,2]) * Permutation([2,1,3])
Permutation([2, 3, 1], 3)

julia> Permutation([2,1,3]) * Permutation([1,3,2])
Permutation([3, 1, 2], 3)

source

Base.:^ — Method.

^(perm ::Permutation, pow ::Integer)

Exponentiate the permutation.

Examples

julia> using ExactDiagonalization

julia> Permutation([2,3,4,1])^2
Permutation([3, 4, 1, 2], 2)

source

ExactDiagonalization.splitblock — Method.

splitblock

Split n into b blocks.

Arguments

n ::Integer: the number of elements to split.

b ::Integer: the number of blocks.

source