API

ExactDiagonalization.HilbertSpace — Type

HilbertSpace{QN}

Abstract Hilbert space with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> spin_site = Site{Int64}([State{Int64}("Up", +1), State{Int64}("Dn", -1)])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

julia> hs = HilbertSpace{Int64}([spin_site, spin_site])
HilbertSpace{Int64}(Site{Int64}[Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)]), Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])], [1, 1], [0, 1, 2])

ExactDiagonalization.Site — Type

Site{QN}

A site with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> up = State{Int}("Up", 1); dn = State("Dn", -1);

julia> Site([up, dn])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

ExactDiagonalization.SparseState — Type

struct SparseState{Scalar<:Number, BR}

Represents a row vector. Free.

ExactDiagonalization.State — Type

State{QN}

State with quantum number type QN.

Examples

julia> using ExactDiagonalization, StaticArrays

julia> up = State{Int}("Up", 1)
State{Int64}("Up", 1)

julia> State("Dn", SVector{2, Int}([-1, 1]))
State{SArray{Tuple{2},Int64,1,2}}("Dn", [-1, 1])

ExactDiagonalization.apply! — Method

apply!

Apply operator to psi and add it to out.

ExactDiagonalization.apply! — Method

apply!(out, opr, state)

Perform out += opr * state. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Call apply_serial! if Threads.nthreads() == 1, and apply_parallel! if greater.

Arguments

out ::Vector{S1}

state ::AbstractVector{S2}

opr ::AbstractOperatorRepresentation{S}

ExactDiagonalization.apply! — Method

apply!(out, opr, state)

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Call apply_serial! if Threads.nthreads() == 1, and apply_parallel! if greater.

Arguments

out ::Vector{S1}

opr ::OperatorRepresentation{HSR, O}

state ::AbstractVector{S2}

ExactDiagonalization.apply_parallel! — Method

apply_parallel!(out, state, opr; range=1:size(opr, 1))

Perform out += state * opr. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Multi-threaded version.

Arguments

out ::Vector{S1}

state ::AbstractVector{S2}

opr ::OperatorRepresentation{HSR, O}

range ::AbstractVector{<:Integer}=1:dimension(opr.hilbert_space_representation)

ExactDiagonalization.apply_parallel! — Method

apply_parallel!(out, opr, state; range=1:size(opr, 2))

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Multi-threaded version.

Arguments

out ::Vector{S1}

opr ::AbstractOperatorRepresentation{S}

state ::AbstractVector{S2}

range ::AbstractVector{<:Integer}=1:size(opr, 2)

ExactDiagonalization.apply_serial! — Method

apply_serial!(out, state, opr; range=1:size(opr, 1))

Perform out += state * opr. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Single-threaded version.

Arguments

out ::Vector{S1}

state ::AbstractVector{S2}

opr ::AbstractOperatorRepresentation{S}

range ::AbstractVector{<:Integer}=1:size(opr, 1)

ExactDiagonalization.apply_serial! — Method

apply_serial!(out, opr, state; range=1:size(opr, 2))

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Single-threaded version.

Arguments

out ::Vector{S1}

opr ::AbstractOperatorRepresentation{S}

state ::AbstractVector{S2}

range ::AbstractVector{<:Integer}=1:size(opr, 1)

ExactDiagonalization.bitwidth — Method

Total number of bits

julia> using ExactDiagonalization

julia> spin_site = Site{Int64}([State{Int64}("Up", +1), State{Int64}("Dn", -1)])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

julia> hs = HilbertSpace{Int64}([spin_site, spin_site, spin_site,])
HilbertSpace{Int64}(Site{Int64}[Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)]), Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)]), Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])], [1, 1, 1], [0, 1, 2, 3])

julia> bitwidth(hs)
3

ExactDiagonalization.bitwidth — Method

bitwidth(site ::Site)

Number of bits necessary to represent the states of the given site.

ExactDiagonalization.extract — Method

Convert binary representation to an array of indices (of states)

Examples ≡≡≡≡≡≡≡≡≡≡

ExactDiagonalization.get_quantum_number — Method

get_quantum_number

ExactDiagonalization.get_row_iterator — Method

May contain duplicates

ExactDiagonalization.get_state — Method

get_state(site ::Site{QN}, binrep ::BR) where {QN, BR<:Unsigned}

Returns the state of site represented by the bits binrep.

ExactDiagonalization.qntype — Method

qntype

ExactDiagonalization.quantum_number_sectors — Method

quantum_number_sectors

ExactDiagonalization.represent — Method

represent(hs; BR=UInt)

Make a HilbertSpaceRepresentation with all the basis vectors of the specified HilbertSpaceSector.

Arguments

hs ::AbstractHilbertSpace

BR ::DataType=UInt: Binary representation type

ExactDiagonalization.represent — Method

represent(hs, basis_list)

Make a HilbertSpaceRepresentation with the provided list of basis vectors

Arguments

hs ::AbstractHilbertSpace

basis_list ::AbstractVector{BR}

ExactDiagonalization.represent_dict — Method

represent_dict(hs; BR=UInt)

Make a HilbertSpaceRepresentation with all the basis vectors of the specified HilbertSpace.

Arguments

hs ::AbstractHilbertSpace

BR ::DataType=UInt: Binary representation type

ExactDiagonalization.represent_dict — Method

represent_dict(hs, basis_list)

Make a HilbertSpaceRepresentation with the provided list of basis vectors using Dict

Arguments

hs ::HilbertSpace{QN}: Abstract Hilbert space

basis_list ::AbstractVector{BR}

ExactDiagonalization.symmetry_reduce — Method

ExactDiagonalization.symmetry_reduce_serial — Method

symmetry_reduce_serial(hsr, trans_group, frac_momentum; ComplexType=ComplexF64, tol=sqrt(eps(Float64)))

Symmetry-reduce the HilbertSpaceRepresentation using translation group.

TightBindingLattice.dimension — Method

dimension

Dimension of the Concrete Hilbert space, i.e. number of basis vectors.

TightBindingLattice.dimension — Method

dimension(site ::Site)

Hilbert space dimension of a given site ( = number of states).

ExactDiagonalization.splitblock — Method

splitblock

Split n into b blocks.

Arguments

n ::Integer: the number of elements to split.

b ::Integer: the number of blocks.