API

ExactDiagonalization.AbstractHilbertSpaceRepresentation — Type

AbstractHilbertSpaceRepresentation{S}

ExactDiagonalization.AbstractOperatorRepresentation — Type

AbstractOperatorRepresentation{S}

ExactDiagonalization.HilbertSpace — Type

HilbertSpace{QN}

Abstract Hilbert space with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> spin_site = Site{Int64}([State{Int64}("Up", +1), State{Int64}("Dn", -1)])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

julia> hs = HilbertSpace{Int64}([spin_site, spin_site])
HilbertSpace{Int64}(Site{Int64}[Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)]), Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])], [1, 1], [0, 1, 2])

ExactDiagonalization.HilbertSpaceRepresentation — Type

HilbertSpaceRepresentation{HS, BR, DictType}

Fields

hilbert_space :: HS
basis_list    :: Vector{BR}
basis_lookup  :: DictType

ExactDiagonalization.HilbertSpaceSector — Type

HilbertSpaceSector{QN}

Hilbert space sector.

ExactDiagonalization.NullOperator — Type

NullOperator

A null operator, i.e. 0.

ExactDiagonalization.OperatorRepresentation — Type

OperatorRepresentation{HSR, S, O}

Operator representation of given operator of type O.

ExactDiagonalization.PureOperator — Type

PureOperator{Scalar, BR}

Represents an operator $α (P₁ ⊗ P₂ ⊗ … ⊗ Pₙ)$ where $Pᵢ$ is either identity (when bitmask is set to zero), or projection $|rᵢ⟩⟨cᵢ|$ (when bitmask is set to one).

Fields

bitmask   :: BR
bitrow    :: BR
bitcol    :: BR
amplitude :: Scalar

ExactDiagonalization.ReducedHilbertSpaceRepresentation — Type

ReducedHilbertSpaceRepresentation{HSR, BR, C}

Representation of the symmetry-reduced hilbert space. Currently only supports Translation group (i.e. Abelian group). ```

ExactDiagonalization.ReducedOperatorRepresentation — Type

ReducedOperatorRepresentation{RHSR, O, S, BR}

Representation of an operator of type O in the symmetry-reduced hilbert space representation of type RHSR.

ExactDiagonalization.Site — Type

Site{QN}

A site with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> up = State{Int}("Up", 1); dn = State("Dn", -1);

julia> Site([up, dn])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

ExactDiagonalization.SparseState — Type

struct SparseState{Scalar<:Number, BR}

Represents a row vector. Free.

ExactDiagonalization.State — Type

State{QN}

State with quantum number type QN.

Examples

julia> using ExactDiagonalization, StaticArrays

julia> up = State{Int}("Up", 1)
State{Int64}("Up", 1)

julia> State("Dn", SVector{2, Int}([-1, 1]))
State{SArray{Tuple{2},Int64,1,2}}("Dn", [-1, 1])

ExactDiagonalization.SumOperator — Type

SumOperator{Scalar, BR}

Represents a sum of pure operators.

Members

terms::Vector{PureOperator{Scalar,BR}}

ExactDiagonalization.apply! — Method

apply!

Apply operator to psi and add it to out.

ExactDiagonalization.apply! — Method

apply!(out, opr, state)

Perform out += opr * state. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Call apply_serial! if Threads.nthreads() == 1, and apply_parallel! if greater.

Arguments

out ::Vector{S1}

state ::AbstractVector{S2}

opr ::AbstractOperatorRepresentation{S}

ExactDiagonalization.apply! — Method

apply!(out, opr, state)

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Call apply_serial! if Threads.nthreads() == 1, and apply_parallel! if greater.

Arguments

out ::Vector{S1}

opr ::OperatorRepresentation{HSR, O}

state ::AbstractVector{S2}

ExactDiagonalization.apply_parallel! — Method

apply_parallel!(out, state, opr; range=1:size(opr, 1))

Perform out += state * opr. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Multi-threaded version.

Arguments

out ::Vector{S1}

state ::AbstractVector{S2}

opr ::OperatorRepresentation{HSR, O}

range ::AbstractVector{<:Integer}=1:dimension(opr.hilbert_space_representation)

ExactDiagonalization.apply_parallel! — Method

apply_parallel!(out, opr, state; range=1:size(opr, 2))

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Multi-threaded version.

Arguments

out ::Vector{S1}

opr ::AbstractOperatorRepresentation{S}

state ::AbstractVector{S2}

range ::AbstractVector{<:Integer}=1:size(opr, 2)

ExactDiagonalization.apply_serial! — Method

apply_serial!(out, state, opr; range=1:size(opr, 1))

Perform out += state * opr. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Single-threaded version.

Arguments

out ::Vector{S1}

state ::AbstractVector{S2}

opr ::AbstractOperatorRepresentation{S}

range ::AbstractVector{<:Integer}=1:size(opr, 1)

ExactDiagonalization.apply_serial! — Method

apply_serial!(out, opr, state; range=1:size(opr, 2))

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Single-threaded version.

Arguments

out ::Vector{S1}

opr ::AbstractOperatorRepresentation{S}

state ::AbstractVector{S2}

range ::AbstractVector{<:Integer}=1:size(opr, 1)

ExactDiagonalization.basespace — Method

basespace(hs)

Get the base space of the HilbertSpace hs, which is itself.

ExactDiagonalization.basespace — Method

basespace(hss)

Get the base space of the HilbertSpaceSector, which is its parent HilbertSpace (with no quantum number restriction).

ExactDiagonalization.bitwidth — Method

Total number of bits

julia> using ExactDiagonalization

julia> spin_site = Site{Int64}([State{Int64}("Up", +1), State{Int64}("Dn", -1)])
Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])

julia> hs = HilbertSpace{Int64}([spin_site, spin_site, spin_site,])
HilbertSpace{Int64}(Site{Int64}[Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)]), Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)]), Site{Int64}(State{Int64}[State{Int64}("Up", 1), State{Int64}("Dn", -1)])], [1, 1, 1], [0, 1, 2, 3])

julia> bitwidth(hs)
3

ExactDiagonalization.bitwidth — Method

bitwidth(site ::Site)

Number of bits necessary to represent the states of the given site.

ExactDiagonalization.compress — Method

Convert an array of indices (of states) to binary representation

ExactDiagonalization.compress — Method

compress(site, state_index, binary_type=UInt) :: binary_type

Get binary representation of the state specified by state_index. Check bounds 1 <= state_index <= dimension(site), and returns binary representation of state_index-1.

Arguments

site::Site

state_index::Integer

binary_type

ExactDiagonalization.extract — Method

Convert binary representation to an array of indices (of states)

Examples ≡≡≡≡≡≡≡≡≡≡

ExactDiagonalization.get_column_iterator — Method

get_column_iterator(op, bc ::BR)

Returns an iterator over the elements of the column corresponding to bit representation bc.

ExactDiagonalization.get_column_iterator — Method

get_column_iterator(opr, icol)

Returns an iterator which generates a list of elements of the column icol. Each element is represented as (irow, amplitude). May contain duplicates and invalid elements. Invalid elements are represented as (-1, amplitude).

ExactDiagonalization.get_element — Method

get_element(opr, irow, icol)

ExactDiagonalization.get_quantum_number — Method

get_quantum_number

ExactDiagonalization.get_quantum_number — Method

get_quantum_number(site, state_index)

Gets the quantum number of state specified by state_index.

ExactDiagonalization.get_row_iterator — Method

get_row_iterator(op, br ::BR)

Returns an iterator over the elements of the row corresponding to bit representation br.

ExactDiagonalization.get_row_iterator — Method

get_row_iterator(opr, irow)

Returns an iterator which generates a list of elements of the row irow. Each element is represented as (icol, amplitude). May contain duplicates and invalid elements. Invalid elements are represented as (-1, amplitude).

ExactDiagonalization.get_state — Method

get_state(site ::Site{QN}, binrep ::BR) where {QN, BR<:Unsigned}

Returns the state of site represented by the bits binrep.

ExactDiagonalization.get_state_index — Method

get_state_index(site, binrep)

Gets the state index of the binary representation. Returns Int(binrep+1).

ExactDiagonalization.pure_operator — Method

pure_operator

Creates a pure operator where projection is at one of the sites.

Arguments

hilbert_space :: HilbertSpace
isite         :: Integer
istate_row    :: Integer
istate_col    :: Integer
amplitude     :: S=1
binary_type   :: Type{BR}=UInt

ExactDiagonalization.qntype — Method

qntype(arg ::Type{HilbertSpaceSector{QN}})

Returns the quantum number type of the given hilbert space sector type.

ExactDiagonalization.qntype — Method

qntype(arg ::Type{HilbertSpace{QN}})

Returns the quantum number type of the given hilbert space type.

ExactDiagonalization.qntype — Method

qntype(::Type{Site{QN}})

Returns the quantum number type of the given site type.

ExactDiagonalization.qntype — Method

qntype(::Type{State{QN}})

Returns the quantum number type of the given state type.

ExactDiagonalization.quantum_number_sectors — Method

quantum_number_sectors

ExactDiagonalization.quantum_number_sectors — Method

quantum_number_sectors(site :: Site{QN}) :: Vector{QN}

Gets a list of possible quantum numbers as a sorted vector of QN.

ExactDiagonalization.represent_array — Method

represent_array(hs, binary_type=UInt)

Make a HilbertSpaceRepresentation with all the basis vectors of the specified HilbertSpaceSector using FrozenSortedArrayIndex{BR}.

Arguments

hs :: AbstractHilbertSpace

binary_type = UInt: Binary representation type

ExactDiagonalization.represent_array — Method

represent_array(hs, basis_list)

Make a HilbertSpaceRepresentation with the provided list of basis vectors using Dict{BR, Int}.

Arguments

hs :: AbstractHilbertSpace

basis_list :: AbstractVector{BR}

ExactDiagonalization.represent_dict — Method

represent_dict(hs, binary_type=UInt)

Make a HilbertSpaceRepresentation with all the basis vectors of the specified HilbertSpace.

Arguments

hs :: AbstractHilbertSpace

binary_type = UInt: Binary representation type

ExactDiagonalization.represent_dict — Method

represent_dict(hs, basis_list)

Make a HilbertSpaceRepresentation with the provided list of basis vectors using Dict.

Arguments

hs ::HilbertSpace{QN}: Abstract Hilbert space

basis_list ::AbstractVector{BR}

ExactDiagonalization.scalartype — Method

scalartype(lhs::Type{<:AbstractOperator{S}})

Returns the scalar type of the given AbstractOperator.

ExactDiagonalization.scalartype — Method

scalartype(arg ::Type{HilbertSpaceSector{QN}})

Returns the scalar type of the given hilbert space sector type. For HilbertSpaceSector{QN}, it is always Bool.

ExactDiagonalization.scalartype — Method

scalartype(arg ::Type{HilbertSpace{QN}})

Returns the scalar type of the given hilbert space type. For HilbertSpace{QN}, it is always Bool.

ExactDiagonalization.simplify — Method

simplify

Simplify the given operator.

ExactDiagonalization.symmetry_reduce — Method

symmetry_reduce(hsr, trans_group, frac_momentum, complex_type=ComplexF64, tol=sqrt(eps(Float64)))

Symmetry-reduce the HilbertSpaceRepresentation using translation group.

ExactDiagonalization.symmetry_reduce — Method

ExactDiagonalization.symmetry_reduce_parallel — Method

symmetry_reduce_parallel(hsr, trans_group, frac_momentum, complex_type=ComplexF64, tol=sqrt(eps(Float64)))

Symmetry-reduce the HilbertSpaceRepresentation using translation group (multi-threaded).

ExactDiagonalization.symmetry_reduce_serial — Method

symmetry_reduce_serial(hsr, trans_group, frac_momentum, complex_type=ComplexF64, tol=sqrt(eps(Float64)))

Symmetry-reduce the HilbertSpaceRepresentation using translation group (single threaded).

TightBindingLattice.dimension — Method

dimension

Dimension of the Concrete Hilbert space, i.e. number of basis vectors.

TightBindingLattice.dimension — Method

dimension(site ::Site)

Hilbert space dimension of a given site (= number of states).

TightBindingLattice.dimension — Method

dimension(arg ::ReducedHilbertSpaceRepresentation{HSR, BR, C}) -> Int

Dimension of the given reduced hilbert space representation, i.e. number of basis elements.

Base.valtype — Method

valtype(lhs::Type{<:AbstractOperator{S}})

Returns the valtype (scalar type) of the given AbstractOperator.

Base.valtype — Method

valtype(arg ::Type{HilbertSpaceSector{QN}})

Returns the valtype (scalar type) of the given hilbert space sector type. For HilbertSpaceSector{QN}, it is always Bool.

Base.valtype — Method

valtype(arg ::Type{HilbertSpace{QN}})

Returns the valtype (scalar type) of the given hilbert space type.

ExactDiagonalization.splitblock — Method

splitblock

Split n into b blocks.

Arguments

n ::Integer: the number of elements to split.

b ::Integer: the number of blocks.