API

ExactDiagonalization.AbstractHilbertSpaceRepresentation — Type

AbstractHilbertSpaceRepresentation{S}

ExactDiagonalization.AbstractOperator — Type

AbstractOperator{S<:Number}

Represent an abstract operator in Hilbert space.

ExactDiagonalization.AbstractOperatorRepresentation — Type

AbstractOperatorRepresentation{S}

ExactDiagonalization.FrozenSortedArrayIndex — Type

FrozenSortedArrayIndex(items)

An immutable sorted array, with index lookup using binary search.

ExactDiagonalization.HilbertSpace — Type

HilbertSpace{QN}

Abstract Hilbert space with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> spin_site = Site([State("Up", +1), State("Dn", -1)]);

julia> hs = HilbertSpace([spin_site, spin_site])
HilbertSpace{Tuple{Int64}}(Site{Tuple{Int64}}[Site{Tuple{Int64}}(State{Tuple{Int64}}[State{Tuple{Int64}}("Up", (1,)), State{Tuple{Int64}}("Dn", (-1,))]), Site{Tuple{Int64}}(State{Tuple{Int64}}[State{Tuple{Int64}}("Up", (1,)), State{Tuple{Int64}}("Dn", (-1,))])], [1, 1], [0, 1, 2])

ExactDiagonalization.HilbertSpaceRepresentation — Type

HilbertSpaceRepresentation{HS, BR, DictType}

Fields

hilbert_space :: HS
basis_list    :: Vector{BR}
basis_lookup  :: DictType

ExactDiagonalization.HilbertSpaceSector — Type

HilbertSpaceSector{QN}

Hilbert space sector.

ExactDiagonalization.IntegerModulo — Type

IntegerModulo{N}

Implement Zₙ.

ExactDiagonalization.NullOperator — Type

NullOperator

A null operator, i.e. 0.

ExactDiagonalization.OperatorRepresentation — Type

OperatorRepresentation{HSR, S, O}

Operator representation of given operator of type O.

ExactDiagonalization.PureOperator — Type

PureOperator{Scalar, BR}

Represents an operator $α (P₁ ⊗ P₂ ⊗ … ⊗ Pₙ)$ where $Pᵢ$ is either identity (when bitmask is set to zero), or projection $|rᵢ⟩⟨cᵢ|$ (when bitmask is set to one).

See also: pure_operator

Fields

bitmask   :: BR
bitrow    :: BR
bitcol    :: BR
amplitude :: Scalar

ExactDiagonalization.ReducedHilbertSpaceRepresentation — Type

ReducedHilbertSpaceRepresentation{HSR, BR, C}

Representation of the symmetry-reduced hilbert space. Currently only supports Translation group (i.e. Abelian group). ```

ExactDiagonalization.ReducedOperatorRepresentation — Type

ReducedOperatorRepresentation{RHSR, O, S, BR}

Representation of an operator of type O in the symmetry-reduced hilbert space representation of type RHSR.

ExactDiagonalization.Site — Type

Site{QN}

A site with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> site = Site([State("Up", 1), State("Dn", -1)]);

ExactDiagonalization.SparseState — Type

struct SparseState{Scalar<:Number, BR}

Represents a vector in unrestricted Hilbert space.

ExactDiagonalization.SparseState — Method

SparseState(hsrep, state_rep, tol=√eps(Float64))

Make a SparseState from a representation

ExactDiagonalization.State — Type

State{QN}

State with quantum number type QN.

Examples

julia> using ExactDiagonalization

julia> up = State("Up", 1)
State{Tuple{Int64}}("Up", (1,))

julia> State("Dn", (-1, 1))
State{Tuple{Int64,Int64}}("Dn", (-1, 1))

ExactDiagonalization.SumOperator — Type

SumOperator{Scalar, BR}

Represents a sum of pure operators.

Fields

terms::Vector{PureOperator{Scalar,BR}}

Base.valtype — Method

valtype(::Type{HilbertSpace{QN}})

Returns the valtype (scalar type) of the given hilbert space type.

Base.valtype — Method

valtype(arg::Type{HilbertSpaceSector{HS, QN}})

Returns the valtype (scalar type) of the given hilbert space sector type. For HilbertSpaceSector{QN}, it is always Bool.

Base.valtype — Method

valtype(lhs::Type{<:AbstractOperator{S}})

Returns the valtype (scalar type) of the given AbstractOperator.

ExactDiagonalization.apply! — Method

apply!(out, nullop, psi)

Apply operator to psi and add it to out.

ExactDiagonalization.apply! — Method

apply!(out, opr, state)

Perform out += opr * state. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Call apply_serial! if Threads.nthreads() == 1, and apply_parallel! otherwise.

ExactDiagonalization.apply! — Method

apply!(out, opr, state)

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Call apply_serial! if Threads.nthreads() == 1, and apply_parallel! otherwise.

ExactDiagonalization.apply_parallel! — Method

apply_parallel!(out, state, opr)

Perform out += state * opr. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Multi-threaded version.

ExactDiagonalization.apply_parallel! — Method

apply_parallel!(out, opr, state)

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Multi-threaded version.

ExactDiagonalization.apply_serial! — Method

apply_serial!(out, state, opr)

Perform out += state * opr. Apply the operator representation opr to the row vector state and add it to the row vector out. Return sum of errors and sum of error-squared. Single-threaded version.

ExactDiagonalization.apply_serial! — Method

apply_serial!(out, opr, state)

Perform out += opr * state. Apply the operator representation opr to the column vector state and add it to the column vector out. Return sum of errors and sum of error-squared. Single-threaded version.

ExactDiagonalization.basespace — Method

basespace(hs)

Get the base space of the HilbertSpace hs, which is itself.

ExactDiagonalization.basespace — Method

basespace(hss)

Get the base space of the HilbertSpaceSector, which is its parent HilbertSpace (with no quantum number restriction).

ExactDiagonalization.bitwidth — Method

bitwidth(hss::HilbertSpaceSector, args...;kwargs...)

Call bitwidth with basespace of hss.

ExactDiagonalization.bitwidth — Method

Total number of bits

julia> using ExactDiagonalization

julia> spin_site = Site([State("Up", +1), State("Dn", -1)])
Site{Tuple{Int64}}(State{Tuple{Int64}}[State{Tuple{Int64}}("Up", (1,)), State{Tuple{Int64}}("Dn", (-1,))])

julia> hs = HilbertSpace([spin_site, spin_site, spin_site,])
HilbertSpace{Tuple{Int64}}(Site{Tuple{Int64}}[Site{Tuple{Int64}}(State{Tuple{Int64}}[State{Tuple{Int64}}("Up", (1,)), State{Tuple{Int64}}("Dn", (-1,))]), Site{Tuple{Int64}}(State{Tuple{Int64}}[State{Tuple{Int64}}("Up", (1,)), State{Tuple{Int64}}("Dn", (-1,))]), Site{Tuple{Int64}}(State{Tuple{Int64}}[State{Tuple{Int64}}("Up", (1,)), State{Tuple{Int64}}("Dn", (-1,))])], [1, 1, 1], [0, 1, 2, 3])

julia> bitwidth(hs)
3

ExactDiagonalization.bitwidth — Method

bitwidth(site)

Number of bits necessary to represent the states of the given site.

ExactDiagonalization.compress — Method

compress(hss::HilbertSpaceSector, args...;kwargs...)

Call compress with basespace of hss.

ExactDiagonalization.compress — Method

compress(bitwidths, data, BR)

Compress data array into a binary integer of type BR.

ExactDiagonalization.compress — Method

compress(hs, indexarray::CartesianIndex, binary_type=UInt)

Convert a cartesian index (a of state) to its binary representation

Examples

julia> using ExactDiagonalization

julia> spin_site = Site([State("Up", +1), State("Dn", -1)]);

julia> hs = HilbertSpace([spin_site, spin_site]);

julia> compress(hs, CartesianIndex(2,2))
0x0000000000000003

ExactDiagonalization.compress — Method

compress(site, state_index, binary_type=UInt) -> binary_type

Get binary representation of the state specified by state_index. Check bounds 1 <= state_index <= dimension(site), and returns binary representation of state_index-1.

ExactDiagonalization.extract — Method

extract(hss::HilbertSpaceSector, args...;kwargs...)

Call extract with basespace of hss.

ExactDiagonalization.extract — Method

Convert binary representation to an array of indices (of states)

Examples

julia> using ExactDiagonalization

julia> spin_site = Site([State("Up", +1), State("Dn", -1)]);

julia> hs = HilbertSpace([spin_site, spin_site]);

julia> extract(hs, 0x03)
CartesianIndex(2, 2)

ExactDiagonalization.get_bitmask — Method

get_bitmask(hss::HilbertSpaceSector, args...;kwargs...)

Call get_bitmask with basespace of hss.

ExactDiagonalization.get_column_iterator — Method

get_column_iterator(op, bcol)

Returns an iterator over the elements of the column corresponding to bit representation bc.

ExactDiagonalization.get_column_iterator — Method

get_column_iterator(opr, icol)

Returns an iterator which generates a list of elements of the column icol. Each element is represented as (irow, amplitude). May contain duplicates and invalid elements. Invalid elements are represented as (-1, amplitude).

ExactDiagonalization.get_element — Method

get_element(opr, irow, icol)

ExactDiagonalization.get_quantum_number — Method

get_quantum_number(hss::HilbertSpaceSector, args...;kwargs...)

Call get_quantum_number with basespace of hss.

ExactDiagonalization.get_quantum_number — Method

get_quantum_number

ExactDiagonalization.get_quantum_number — Method

get_quantum_number(site, state_index)

Gets the quantum number of state specified by state_index.

ExactDiagonalization.get_row_iterator — Method

get_row_iterator(op, br)

Returns an iterator over the elements of the row corresponding to bit representation br.

ExactDiagonalization.get_row_iterator — Method

get_row_iterator(ropr::ROR, irow_r::Integer)

Get the row iterator for the reduced operator representation

ExactDiagonalization.get_row_iterator — Method

get_row_iterator(opr, irow)

Returns an iterator which generates a list of elements of the row irow. Each element is represented as (icol, amplitude). May contain duplicates and invalid elements. Invalid elements are represented as (-1, amplitude).

ExactDiagonalization.get_state — Method

get_state(hss::HilbertSpaceSector, args...;kwargs...)

Call get_state with basespace of hss.

ExactDiagonalization.get_state — Method

get_state(hs, binrep, isite)

Get the local state at site isite for the basis state represented by binrep. Returns an object of type State

ExactDiagonalization.get_state — Method

get_state(site, binrep) where {QN, BR<:Unsigned}

Returns the state of site represented by the bits binrep.

ExactDiagonalization.get_state_index — Method

get_state_index(hss::HilbertSpaceSector, args...;kwargs...)

Call get_state_index with basespace of hss.

ExactDiagonalization.get_state_index — Method

get_state_index(hs, binrep, isite)

Get the index of the local state at site isite for the basis state represented by binrep.

ExactDiagonalization.get_state_index — Method

get_state_index(site, binrep)

Gets the state index of the binary representation. Returns Int(binrep+1).

ExactDiagonalization.hs_get_basis_list — Method

hs_get_basis_list(hss, binary_type=UInt)

Generate a basis for the HilbertSpaceSector.

ExactDiagonalization.pure_operator — Method

pure_operator(hilbert_space, isite, istate_row, istate_col, amplitude=1, binary_type=UInt)

Creates a pure operator where projection is at one of the sites.

ExactDiagonalization.qntype — Method

qntype(arg::Type{HilbertSpaceSector{QN}})

Returns the quantum number type of the given hilbert space sector type.

ExactDiagonalization.qntype — Method

qntype(::Type{HilbertSpace{QN}})

Returns the quantum number type of the given hilbert space type.

ExactDiagonalization.qntype — Method

qntype(::Type{Site{QN}})

Returns the quantum number type of the given site type.

ExactDiagonalization.qntype — Method

qntype(::Type{State{QN}})

Returns the quantum number type of the given state type.

ExactDiagonalization.quantum_number_sectors — Method

quantum_number_sectors(hss::HilbertSpaceSector, args...;kwargs...)

Call quantum_number_sectors with basespace of hss.

ExactDiagonalization.quantum_number_sectors — Method

quantum_number_sectors

ExactDiagonalization.quantum_number_sectors — Method

quantum_number_sectors(site) -> Vector{QN}

Gets a list of possible quantum numbers as a sorted vector of QN.

ExactDiagonalization.represent — Method

represent(hs, binary_type=UInt)

Make a HilbertSpaceRepresentation with all the basis vectors of the specified HilbertSpace. This function defaults to represent_array.

ExactDiagonalization.represent — Method

represent(hs, basis_list)

Make a HilbertSpaceRepresentation with the provided list of basis vectors. This defaults to represent_array.

ExactDiagonalization.represent — Method

represent(hilbert_space_representation, operator)

Create an OperatorRepresentation of the operator in the hilbert_space_representation.

ExactDiagonalization.represent_array — Method

represent_array(hs, binary_type=UInt)

Make a HilbertSpaceRepresentation with all the basis vectors of the specified HilbertSpace using FrozenSortedArrayIndex.

ExactDiagonalization.represent_array — Method

represent_array(hs, basis_list)

Make a HilbertSpaceRepresentation with the provided list of basis vectors using FrozenSortedArrayIndex.

ExactDiagonalization.represent_dict — Method

represent_dict(hs, binary_type=UInt)

Make a HilbertSpaceRepresentation with the provided list of basis vectors using Dict{binary_type, Int}.

ExactDiagonalization.represent_dict — Method

represent_dict(hs, basis_list)

Make a HilbertSpaceRepresentation with the provided list of basis vectors using Dict.

ExactDiagonalization.scalartype — Method

scalartype(::Type{HilbertSpace{QN}})

Returns the scalar type of the given hilbert space type. For HilbertSpace{QN}, it is always Bool.

ExactDiagonalization.scalartype — Method

scalartype([state or type of state])

Return the scalar type of the state.

ExactDiagonalization.scalartype — Method

scalartype(arg::Type{HilbertSpaceSector{HS, QN}})

Returns the scalar type of the given hilbert space sector type. For HilbertSpaceSector{QN}, it is always Bool.

ExactDiagonalization.scalartype — Method

scalartype(lhs::Type{<:AbstractOperator{S}})

Returns the scalar type of the given AbstractOperator.

ExactDiagonalization.simplify — Method

simplify

Simplify the given operator.

ExactDiagonalization.splitblock — Method

splitblock

Split n into b blocks.

ExactDiagonalization.splitrange — Method

splitrange

ExactDiagonalization.symmetry_reduce! — Method

symmetry_reduce!(out, rhsr, largevector)

Adds and not overwrites.

ExactDiagonalization.symmetry_reduce — Method

symmetry_reduce(hsr, lattice, symmetry_irrep_component, complex_type=ComplexF64, tol=√ϵ)

Symmetry-reduce the HilbertSpaceRepresentation using translation group.

ExactDiagonalization.symmetry_reduce — Method

symmetry_reduce(rhsr, large_vector)

Reduce a large vector into the reduced hilbert space representation. Simply throw away components that don't fit.

ExactDiagonalization.symmetry_reduce_parallel — Method

symmetry_reduce_parallel(hsr, trans_group, frac_momentum, complex_type=ComplexF64, tol=√ϵ)

Symmetry-reduce the HilbertSpaceRepresentation using translation group (multi-threaded).

ExactDiagonalization.symmetry_reduce_parallel — Method

symmetry_reduce_parallel(hsr, trans_group, frac_momentum, complex_type=ComplexF64, tol=√ϵ)

Symmetry-reduce the HilbertSpaceRepresentation using translation group (multi-threaded).

ExactDiagonalization.symmetry_reduce_parallel — Method

symmetry_reduce_parallel(hsr, trans_group, frac_momentum, complex_type=ComplexF64, tol=√ϵ)

Symmetry-reduce the HilbertSpaceRepresentation using translation group (multi-threaded).

ExactDiagonalization.symmetry_reduce_parallel — Method

symmetry_reduce_parallel(hsr, symops_and_amplitudes; tol=√ϵ)

Symmetry-reduce the HilbertSpaceRepresentation using translation group (multi-threaded).

ExactDiagonalization.symmetry_reduce_serial — Method

symmetry_reduce_serial(hsr, trans_group, frac_momentum, complex_type=ComplexF64, tol=√ϵ)

Symmetry-reduce the HilbertSpaceRepresentation using translation group (single threaded).

ExactDiagonalization.symmetry_reduce_serial — Method

symmetry_reduce_serial(hsr, trans_group, frac_momentum, complex_type=ComplexF64, tol=√ϵ)

Symmetry-reduce the HilbertSpaceRepresentation using translation group (single threaded).

ExactDiagonalization.symmetry_reduce_serial — Method

symmetry_reduce_serial(hsr, trans_group, frac_momentum, complex_type=ComplexF64, tol=√ϵ)

Symmetry-reduce the HilbertSpaceRepresentation using translation group (single threaded).

ExactDiagonalization.symmetry_reduce_serial — Method

symmetry_reduce_serial

The irreps have to follow certain order:

symmetry_reduce_serial(
    hilbert_space_representation,
)

ExactDiagonalization.update — Method

update(hss::HilbertSpaceSector, args...;kwargs...)

Call update with basespace of hss.

ExactDiagonalization.update — Method

update(hs, binrep, isite, new_state_index)

Update the binary representation of a basis state by changing the state at site isite to a new local state specified by new_state_index.

LatticeTools.dimension — Method

dimension

Dimension of the Concrete Hilbert space, i.e. number of basis vectors.

LatticeTools.dimension — Method

dimension(arg::ReducedHilbertSpaceRepresentation{HSR, BR, C}) -> Int

Dimension of the given reduced hilbert space representation, i.e. number of basis elements.

LatticeTools.dimension — Method

dimension(site)

Hilbert space dimension of a given site (= number of states).